Moved tests needing MDL to mdl_test.go
