+++ ddl/db_partition_test.go
@@ -2513,91 +2513,6 @@ func TestExchangePartitionTableCompatiable(t *testing.T) {
 	require.NoError(t, err)
 }
 
-func TestExchangePartitionMultiTable(t *testing.T) {
-	logutil.BgLogger().Info("mdl related variable status before bootstrap", zap.Bool("EnableMDL", variable.EnableMDL.Load()), zap.Bool("EnableConcurrentDDL", variable.EnableConcurrentDDL.Load()))
-	store := testkit.CreateMockStore(t)
-	tk1 := testkit.NewTestKit(t, store)
-	logutil.BgLogger().Info("mdl related variable status after bootstrap", zap.Bool("EnableMDL", variable.EnableMDL.Load()), zap.Bool("EnableConcurrentDDL", variable.EnableConcurrentDDL.Load()))
-
-	dbName := "ExchangeMultiTable"
-	tk1.MustExec(`create schema ` + dbName)
-	tk1.MustExec(`use ` + dbName)
-	tk1.MustExec(`set global tidb_enable_metadata_lock = 'ON'`)
-	tk1.MustExec(`CREATE TABLE t1 (a int)`)
-	tk1.MustExec(`CREATE TABLE t2 (a int)`)
-	tk1.MustExec(`CREATE TABLE tp (a int) partition by hash(a) partitions 3`)
-	tk1.MustExec(`insert into t1 values (0)`)
-	tk1.MustExec(`insert into t2 values (3)`)
-	tk1.MustExec(`insert into tp values (6)`)
-	logutil.BgLogger().Info("mdl related variable status after inserting rows", zap.Bool("EnableMDL", variable.EnableMDL.Load()), zap.Bool("EnableConcurrentDDL", variable.EnableConcurrentDDL.Load()))
-
-	tk2 := testkit.NewTestKit(t, store)
-	tk2.MustExec(`use ` + dbName)
-	tk3 := testkit.NewTestKit(t, store)
-	tk3.MustExec(`use ` + dbName)
-	tk4 := testkit.NewTestKit(t, store)
-	tk4.MustExec(`use ` + dbName)
-	waitFor := func(col int, tableName, s string) {
-		for {
-			tk4 := testkit.NewTestKit(t, store)
-			tk4.MustExec(`use test`)
-			sql := `admin show ddl jobs where db_name = '` + strings.ToLower(dbName) + `' and table_name = '` + tableName + `' and job_type = 'exchange partition'`
-			res := tk4.MustQuery(sql).Rows()
-			if len(res) == 1 && res[0][col] == s {
-				break
-			}
-			logutil.BgLogger().Info("No match", zap.String("sql", sql), zap.String("s", s))
-			sql = `admin show ddl jobs`
-			res = tk4.MustQuery(sql).Rows()
-			for _, row := range res {
-				strs := make([]string, 0, len(row))
-				for _, c := range row {
-					strs = append(strs, c.(string))
-				}
-				logutil.BgLogger().Info("admin show ddl jobs", zap.Strings("row", strs))
-			}
-			time.Sleep(100 * time.Millisecond)
-		}
-	}
-	var wg sync.WaitGroup
-	wg.Add(1)
-	dumpChan := make(chan struct{})
-	defer func() {
-		close(dumpChan)
-		wg.Wait()
-	}()
-	go testkit.DebugDumpOnTimeout(&wg, dumpChan, 20*time.Second)
-	alterChan1 := make(chan error)
-	alterChan2 := make(chan error)
-	tk3.MustExec(`BEGIN`)
-	tk3.MustExec(`insert into t1 values (1)`)
-	tk3.MustExec(`insert into t2 values (2)`)
-	tk3.MustExec(`insert into tp values (3)`)
-	//require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/ddl/exchangePartitionAutoID", `pause`))
-	go func() {
-		alterChan1 <- tk1.ExecToErr(`alter table tp exchange partition p0 with table t1`)
-	}()
-	waitFor(11, "t1", "running")
-	go func() {
-		alterChan2 <- tk2.ExecToErr(`alter table tp exchange partition p0 with table t2`)
-	}()
-	waitFor(11, "t2", "queueing")
-	tk3.MustExec(`rollback`)
-	logutil.BgLogger().Info("rollback done")
-	//require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/ddl/exchangePartitionAutoID"))
-	require.NoError(t, <-alterChan1)
-	logutil.BgLogger().Info("alter1 done")
-	err := <-alterChan2
-	logutil.BgLogger().Info("alter2 done")
-	tk3.MustQuery(`select * from t1`).Check(testkit.Rows("6"))
-	logutil.BgLogger().Info("select t1 done")
-	tk3.MustQuery(`select * from t2`).Check(testkit.Rows("0"))
-	logutil.BgLogger().Info("select t2 done")
-	tk3.MustQuery(`select * from tp`).Check(testkit.Rows("3"))
-	logutil.BgLogger().Info("select tp done")
-	require.NoError(t, err)
-}
-
 func TestExchangePartitionValidation(t *testing.T) {
 	store := testkit.CreateMockStore(t)
 	tk := testkit.NewTestKit(t, store)
